main/requirements.txt
jsonpickle==2.1.0
PyJWT==2.3.0
boto3==1.20.54
